pyproject.toml
[tool.poetry]
name = "langgraph-udemy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^5.13.2"
langchain = "^0.3.12"
langchain-openai = "^0.2.12"
langgraph = "^0.2.59"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
